Catch non-existent directory in pg_config
The configure scripts check pg_config to see if pg_config is set, but not whether --includedir and --libdir actually exist. If those directories are non-existent, the build fails. I added a check to see if those directories exist, and if not, unset the variables so the script will look in the paths specified in the script.

I also added the default path and /usr/local to the search path for the include and library locations in a BigSQL package install.